Speech improvements, added simple help
The speech queue system did not account for lines being longer than the
manyland limit of 20.  They do now.  If a longer string is passed in to be
spoken, it's broken up into multiple lines.

Also added a rudimentary help method.  Typing "help" or ":help" will cause the
player to state  all of the commands available, with a brief line on what they
do.