Adding data to ship view, small fixes
It's now fairly easy to add multiple contents to any location scene using
"planet placer".  Even if that gets too bulky as we add bodies, at can be
simply split into more interactings, as it's just a bunch of lines checking to
see if the circumstance is right for placing certain items.

Along with the planets and stations that are there, I added info blurbs that
get displayed on the screen/windshield.  This is really nice, I'm looking
forward to adding things like animated asteroids, nebulas, really big
perspectives of planets, etc. etc.