Moving code from buttons to interactings
There are several updates script-wise, but the core thing done is that the
functionality of removing prior "TARGET_<location>" states has been moved out
of "SET TARGET" buttons, and into a separate script that those buttons each
calls by triggering a user action.  This means that when a new planet is added,
the existing "SET TARGET" buttons do not need to be updated individually, only
the new "target_reset" script needs the update now.