Added necessary pause to set target button.
After setting the state that causes the target_reset script to run, there needs
to be a pause in this interacting, or it winds up setting the target before it
gets scrubbed, leaving us with no target.  It seems to happen only on first
attempt per world load, but the pause seems to fix it.